backend/pyproject.toml
[tool.poetry]
name = "trading_journal_backend"
version = "0.1.0"
description = "Backend for Trading Journal"
authors = ["Your Name"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = {extras = ["standard"], version = "^0.115.8"}
sqlmodel = "^0.0.22"
pydantic = "^2.10.6"
alembic = "^1.14.1"
pytest = "^8.3.4"
pytest-asyncio = "^0.25.3"
pytest-mock = "^3.14.0"
httpx = "^0.28.1"
aiosqlite = "^0.21.0"
asyncpg = "^0.30.0"
greenlet = "^3.1.1"
pytest-cov = "^6.0.0"
starlette-admin = "^0.14.1"
python-dotenv = "^1.0.1"
authx = "^1.4.1"
bcrypt = "^4.3.0"
fastapi-events = "^0.12.2"
requests = "^2.32.4"
fastapi-cache2 = {extras = ["memcache"], version = "^0.2.2"}

[tool.poetry.dev-dependencies]
uvicorn = "^0.34.0"
black = "^25.1.0"
pre-commit = "^4.1.0"
flake8 = "^7.1.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
